Update renamed imports
